examples/sls-logger-example/requirements.txt
# SLS 日志集成示例的最小依赖
yai-loguru-support[sls]